+++ crates/objc2-exception-helper/src/try_catch.m
@@ -1,15 +1,16 @@
 // Don't include any headers, cross compilation is difficult to set up
 // properly in such situations.
 
-/// We're linking to `libobjc` in build.rs, so this should be available.
+/// Users are linking to `libobjc`, so this should be available.
 ///
 /// See <https://clang.llvm.org/docs/AutomaticReferenceCounting.html#arc-runtime-objc-retain>.
 id objc_retain(id value);
 
-/// Unsure how C name resolution works, so we make sure to version this symbol.
+/// C has no name resolution and will conflict if multiple versions of this
+/// crate is present, so we make sure to version this symbol.
 ///
 /// Return `unsigned char` since it is guaranteed to be `u8` on all platforms.
-unsigned char rust_objc_sys_0_3_try_catch_exception(void (*f)(void *), void *context, id *error) {
+unsigned char objc2_exception_helper_0_2_try_catch_exception(void (*f)(void *), void *context, id *error) {
     @try {
         f(context);
         return 0;